Use amd64 variants of files under darwin arm64

+++ ext/redcloth_scan/extconf.rb
@@ -1,10 +1,11 @@
 require 'mkmf'
 
 host_cpu = RbConfig::CONFIG['host_cpu']
+host_os = RbConfig::CONFIG['host_os']
 source_files = ['redcloth_attributes', 'redcloth_inline', 'redcloth_scan']
 
 source_files.each do |base|
-  if host_cpu =~ /arm64|aarch64/
+  if host_cpu =~ /arm64|aarch64/ && host_os !~ /darwin/
     File.rename("#{base}_arm64.c", "#{base}.c") if File.exist?("#{base}_arm64.c")
     File.delete("#{base}_amd64.c") if File.exist?("#{base}_amd64.c")
   else
